requirements.txt
pandas
python-dotenv
google-generativeai 